format adjustments
