+++ policyengine_canada/parameters/gov/provinces/nu/tax/income/credits/cost_of_living/max_amount.yaml
@@ -1,4 +1,4 @@
-description: Nunavut cost of living credit maximum amount.
+description: Nunavut provides up to this amount in its cost of living credit.
 values:
   2022-01-01: 1_500
 metadata:
