+++ policyengine_canada/parameters/gov/provinces/nu/tax/income/credits/non_refundable.yaml
@@ -1,4 +1,4 @@
-description: Nunavut list of non_refundable tax credits.
+description: Nunavut provides these non-refundable tax credits.
 values:
   2022-01-01:
     - nu_cost_of_living_credit
