--- policyengine_canada/tests/gov/provinces/nu/tax/income/credits/nu_cost_of_living_credit.yaml
@@ -1,36 +1,0 @@
-- name: Deduction over income 
-  period: 2023
-  input:
-    province_code: NU
-    individual_net_income: 1_000
-    canadian_armed_forces_and_personnel_deduction: 1_500
-  output:
-    nu_cost_of_living_credit: 0
-
-- name: Wrong province
-  period: 2023
-  input:
-    province_code: AB
-    individual_net_income: 100_000
-    canadian_armed_forces_and_personnel_deduction: 1_500
-  output:
-    nu_cost_of_living_credit: 0
-
-- name: Max amount 1_500
-  period: 2023
-  input:
-    province_code: NU
-    individual_net_income: 100_000
-    canadian_armed_forces_and_personnel_deduction: 1_500
-  output:
-    nu_cost_of_living_credit: 1_500
-
-
-- name: 2 percent of 1_000
-  period: 2023
-  input:
-    province_code: NU
-    individual_net_income: 1_000
-    canadian_armed_forces_and_personnel_deduction: 0
-  output:
-    nu_cost_of_living_credit: 20
