+++ policyengine_canada/variables/gov/provinces/nu/tax/income/credits/nu_cost_of_living_credit_post_supplement.py
@@ -1,15 +1,15 @@
 from policyengine_canada.model_api import *
 
 
-class nu_cost_of_living_credit_post_supplement(Variable):
+class nu_cost_of_living_credit(Variable):
     value_type = float
     entity = Person
-    label = "Nunavut cost of living credit post supplement"
+    label = "Nunavut cost of living credit"
     unit = CAD
     definition_period = YEAR
     defined_for = ProvinceCode.NU
 
-    def formula(person, period, parameters):
-        credit = person("nu_cost_of_living_credit", period)
-        supplement = person("nu_cost_of_living_credit_supplement", period)
-        return credit + supplement
+    adds = [
+        "nu_cost_of_living_basic_credit",
+        "nu_cost_of_living_credit_supplement",
+    ]
