+++ policyengine_canada/variables/gov/provinces/nu/tax/income/credits/nu_cost_of_living_credit_supplement.py
@@ -4,7 +4,7 @@ from policyengine_canada.model_api import *
 class nu_cost_of_living_credit_supplement(Variable):
     value_type = float
     entity = Person
-    label = "Nunavut cost of living credit supplement"
+    label = "Nunavut cost of living credit supplement for single parents"
     unit = CAD
     definition_period = YEAR
     defined_for = ProvinceCode.NU
@@ -13,10 +13,7 @@ class nu_cost_of_living_credit_supplement(Variable):
         p = parameters(
             period
         ).gov.provinces.nu.tax.income.credits.cost_of_living.supplement_for_single_parents
-        income = person("individual_net_income", period)
-        armed_deduction = person(
-            "canadian_armed_forces_and_personnel_deduction", period
+        adjusted_income = person(
+            "nu_cost_of_living_credit_adjusted_net_income", period
         )
-        adjusted_income = max_(0, income - armed_deduction)
-
         return min_(max_(0, adjusted_income - p.income) * p.rate, p.max_amount)
